Edit tests for new which logic
- Separate resistance tests into separate file
